refactor: single layer functions
