+++ src/views/DesignEditor/utils/tools/DuplicateFunc.tsx
@@ -3,4 +3,5 @@ export const DuplicateFunc = ({ editor, activeObject }: any) => {
   setTimeout(() => {
     editor.objects.update({ name: activeObject.name })
   }, 10)
+  editor.cancelContextMenuRequest()
 }
