+++ src/views/DesignEditor/utils/tools/PasteFunc.tsx
@@ -1,0 +1,4 @@
+export const PasteFunc = ({ editor }: any) => {
+  editor.objects.paste()
+  editor.cancelContextMenuRequest()
+}
